Add full text search for web pages (#8)
* add dev command to scrape and transform html

* update cache dir

* begin extracting scraper logic to pkg

* update dev full text command

* add debug logger

* ignore env file

* Fix redirects for single urls

* add scraper tests

* rewrite to use varargs

* fix test

Well, this is not good. Putting the test at the same test level as the
current other tests causes it to fail.

* handle 404s

* add initial full text processing

* refactor search tui

* WIP

Getting a good experience with sql will, i think, require actually using
sqlites fts facilities. This means "sweeping" changes (anything
involving existing data storage is sweeping right now). So for now it'
sleft todo.

* add an fts table

* update makefile to create linux version

* add dev db command tree

* improve logging

* refact most search to use search result type

* use sqlite for full-text search

* move sql init into sql file, update trigger logic

* fix null pointer bug

* fix bleve index page

* update fulltext search indexing

* add backup command

* remove fts sqlite index

For now, not using sqlite for search. I ran into perf issues with it.
Not sure if related to using modernc/sqlite rather than the native
mattn/sqlite driver. However, bleve is fast so if it can be made to
accomplish our needs i'll stick with it.

* use bleve for FTS, no more sqlite for now